[#318] Use scandir implementation on other platforms
