+++ iceoryx2-pal/posix/src/linux/dirent.rs
@@ -13,64 +13,10 @@
 #![allow(non_camel_case_types)]
 #![allow(clippy::missing_safety_doc)]
 
-use crate::posix::{free, malloc, types::*};
+use crate::{posix::types::*, scandir_impl};
 
 pub unsafe fn scandir(path: *const c_char, namelist: *mut *mut *mut dirent) -> int {
-    let dirfd = opendir(path);
-    if dirfd.is_null() {
-        return -1;
-    }
-
-    *namelist = core::ptr::null_mut::<*mut dirent>();
-    let mut entries = vec![];
-    const DIRENT_SIZE: usize = core::mem::size_of::<dirent>();
-
-    let cleanup = |entries: &Vec<*mut void>, namelist: *mut *mut dirent| {
-        for entry in entries {
-            free((*entry).cast());
-        }
-
-        if !namelist.is_null() {
-            free(namelist.cast());
-        }
-    };
-
-    loop {
-        let dirent_ptr = malloc(DIRENT_SIZE);
-        let result_ptr: *mut *mut dirent = malloc(core::mem::size_of::<*mut dirent>()).cast();
-
-        if readdir_r(dirfd, dirent_ptr.cast(), result_ptr) != 0 {
-            free(result_ptr.cast());
-            free(dirent_ptr);
-            cleanup(&entries, *namelist);
-
-            closedir(dirfd);
-            return -1;
-        }
-
-        if (*result_ptr).is_null() {
-            free(result_ptr.cast());
-            free(dirent_ptr);
-            break;
-        }
-
-        free(result_ptr.cast());
-        entries.push(dirent_ptr);
-    }
-
-    *namelist = malloc(core::mem::size_of::<*mut *mut dirent>() * entries.len()).cast();
-    if (*namelist).is_null() {
-        cleanup(&entries, *namelist);
-        closedir(dirfd);
-        return -1;
-    }
-
-    for (n, entry) in entries.iter().enumerate() {
-        (*namelist).add(n).write((*entry).cast());
-    }
-
-    closedir(dirfd);
-    entries.len() as _
+    scandir_impl(path, namelist)
 }
 
 pub unsafe fn mkdir(pathname: *const c_char, mode: mode_t) -> int {
